Weichen

+++ nameserver/Dijkstra.h
@@ -1,6 +1,6 @@
 #ifndef __DIJKSTRA_H__
 #define __DIJKSTRA_H__
 
-int dijkstra(int srcpID, int (*e)[20], int *dis, int *book, int nNodes, int nLinks);
+int dijkstra(int srcpID, float (*e)[20], float *dis, int *book, int nNodes, int nLinks);
 
 #endif
